chore(form-service, pdf-service): move resolve ref for pdf to form service
